Updated Service layer, Controller pending
Updated the service layer, and have created sub packages for the
interfaces to make it easier to view. Controllers are almost set, though
fine tuning is needed with getall and get one methods, review streams.

+++ src/main/java/itacademy/s5t2/diceGame/domain/Player.java
@@ -15,7 +15,7 @@ import lombok.Data;
 public class Player {
 	
 	@Id
-	private int idPlayer;
+	private long idPlayer;
 	
 	@CreationTimestamp
 	private Date registrationDate;
